Move chip to common file, fix input font tokens
